move test that tests the binary to the tests directory to ensure the binary is built before testing
